requirements.txt
beautifulsoup4
selenium
webdriver_manager
pyyaml
